Refactor OAuth
