+++ src/OAuth/OAuthSignatureMethod_HMAC_SHA1.php
@@ -1,0 +1,46 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: mxgel
+ * Date: 11/14/16
+ * Time: 2:28 AM
+ */
+
+namespace Knox\Pesapal\OAuth;
+
+
+/**
+ * Class OAuthSignatureMethod_HMAC_SHA1
+ *
+ * @package Knox\Pesapal\OAuth
+ */
+class OAuthSignatureMethod_HMAC_SHA1 extends OAuthSignatureMethod {
+    /**
+     * @return string
+     */
+    function get_name() {
+        return "HMAC-SHA1";
+    }
+
+    /**
+     * @param $request
+     * @param $consumer
+     * @param $token
+     *
+     * @return string
+     */
+    public function build_signature($request, $consumer, $token) {
+        $base_string = $request->get_signature_base_string();
+        $request->base_string = $base_string;
+
+        $key_parts = [
+            $consumer->secret,
+            ($token) ? $token->secret : "",
+        ];
+
+        $key_parts = OAuthUtil::urlencode_rfc3986($key_parts);
+        $key = implode('&', $key_parts);
+
+        return base64_encode(hash_hmac('sha1', $base_string, $key, true));
+    }
+}
